cli/function_templates/python/requirements.txt
memphis-functions
